factory + hib fix

+++ src/main/java/ua/com/startup/investments/entities/Investment.java
@@ -52,19 +52,19 @@ public class Investment implements Serializable {
     /**
      * Getters and setters methods by all fields of investment.
      */
-    public long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
-    public int getSum() {
+    public Integer getSum() {
         return sum;
     }
 
-    public void setSum(int sum) {
+    public void setSum(Integer sum) {
         this.sum = sum;
     }
